Module Wordpress.Metadata

Data structures attachable to articles/documents.

Metadata description

Minimum interfaces for describing metadata sets.

Injectable

Describes a data set that can be injected, for example into a template. (Currently, the injection process relay on Mustache)

module type INJECTABLE = sig ... end

Parsable

Describes a data set that can be parsed from, for example, the metadata of a document. Which are usually described using Yaml.

module type PARSABLE = sig ... end

Utility

A collection of public features for building metadata sets.

Date

A rather naive representation of dates.

module Date : sig ... end

Rules

Validating JSON (for Mustache and Yaml) using Applicative Validation.

module Rules : sig ... end

Metadata description

Example of a metadata set. These examples are directly usable but you should write your own!

A single page

This collection of metadata describes a single page, characterised by a title and a description. Both parameters are optional.

module Page : sig ... end

An Article

This collection of metadata describes an article, characterised by a title and a description (optional like for single page). And article-related metadata: article_title, article_description, tags and date.

module Article : sig ... end

A page with a list of article

module Articles : sig ... end